Fix vanilla-tx
This tweaks the feature to support vanilla-tx. The hint on the
Transaction Selector needing altering so that the autoconfig
wasn't deactivated if one of the specific types it needs is not
around, since they are not all always required.

This also refactors out the configuration and hints into proper
classes rather than mixing those in with the rest of the codebase.

Fixes gh-40