Ensuring running without 'remove autoconfig' works
With the recent graal fixes we can now run samples without
the 'remove autoconfig' option. Would you want to? Probably not.
But sometimes the situation can be forced because the agent may bring
in the unfiltered spring.factories. This would affect trying to
use the agent and the feature together (without the feature
being in initialization-only mode).