Tweaking behaviour of functional mode for spring-init
For spring-init we need to walk spring.factories, even in functional
mode, but there is only a need to register hint related information
and not reflective configuration for the configurations or configuration
infrastructure (e.g. conditions). These changes try to get close to that
there is a little way further to go.  functional mode only scans
spring.factories if spring-init is around. Without spring-init it
will presume the app is completely in charge of everything (so there
should be no impact to fu related samples).