Attempt to fix validator sample on Java11
On Java11, without this change, there is an error about
registerNatives() being found in the .class bytes
for Object.class but not found when reflecting on the
class (when running parameter name discovery in Spring).
That is because we were adding Object with public methods
visible via reflection, not declared methods. For some
reason this doesn't matter on Java8. Not sure why right
now but the quick way to get the builds green should
be to open up visibility of those methods again (removing
them was only a recent change). Maybe Spring could be
more lenient in a native-image context.