Workaround for tomcat 9.0.39 problem
In tomcat 9.0.39 they deleted a package, removing a resource bundle
but they ship their own native-image config and didn't remove the
bundle from there. Unusually native-image doesn't seem to allow
these to be missing so it fails.  Tomcat have since fixed the
native-image config but that doesn't help 9.0.39. So this file
acts as a dumb placeholder to meet the need of the config. We
can delete when 9.0.40 comes out I imagine.