Provide hsqldb database support
This commit adds new HSQLDB hints to import required classes
and resources.

It also adds a substitution for hsqldb class FrameworkLogger.
This class attempts to retrieve log configuration at runtime
but cannot find them since the property java.home is not set.

This method does not exists anymore in the next development
version (see https://sourceforge.net/p/hsqldb/svn/6213/) so
this substitution should be removed when the next release will
be integrated in Spring Boot.

Closes gh-303