Fix class resolution issue with split packages
Prior to this commit, the `TypeSystem` would store the same JAR entry
multiple times in the cache. This commit turns the cache entries to
`Set` instance of `List` to avoid storing duplicates.

Prior to this commit, the `TypeSystem` would only consider the app
packages for loading classes, for a given package. This would cause
issues in situations where the app would have a split package situation
between the application classes and other classpath entries.
This would happen with build-time transformation where we ship many
classes in different packages to overcome code visibility limitations.

This commit ensures that the JAR entries are also checked if the class
cannot be found in the app classes.