Refactoring towards build-time-infrastructure
Starting small, refactoring the proxy handling to introduce a collector
for all configuration that can optionally feed to the GraalVM
build system directly. It can also be asked at the end for all
that data and place it on the file system (as would happen with
proper build time only analysis).