Avoid statically instantiating factories
Prior to this commit, the bootstrap contributor generating code for
spring.factories would statically instantiate factories and store them
in a map.
This means that the instantiation order would differ from a regular app
using reflection.

This commit switches the implementation to using `Supplier` in order to
delay the actual factory instantiation when it's needed.