Remove static configuration files and functional mode
This commit is a refactoring that turns the remaining
static configuration files into proper annotation-based hints.

It also removes the functional mode which is not something we
want to maintain as it is. Refining the hints trigger combined
with future build time transformation such as gh-386 should provide
the same functionality implicitly.

The rsocket sample is broken, the allowUnsafeAccess defined in
the hints seems not taken in account.

Closes gh-216
Closes gh-347