Generate reflection data for Sprig factories
Prior to this commit, the build plugin would generate source code for
Spring Factories, including a specific case for factories that are
looked up by their name and later instantiated via reflection.

This was causing issues when building native images since the reflection
metadata was missing for those classes.

This commit adds reflection data on the classes themselves (but not on
the public constructors for now).