Configure native build plugin for testing phase
Prior to this commit, the build-time infrastructure plugin would be run
on the "process-classes" phase, so right after source compilation and
before running the test suite.
This was too early as we'd generate sources and compile them with the
assumptions made on the compile+runtime classpath at that point.

Running tests with this set of sources would fail because test
dependencies would be missing from the picture when bootstrap sources
were generated.

This commit revisits the execution phases for the native build plugin:
* one during the "process-test-classes" phase with the test classpath,
before running the test suite
* another one during the "prepare-package" phase with the
compile+runtime classpath before packaging the application.

Fixes gh-456