Require native-configuration in build plugins
Prior to this commit, configuring the spring-native-maven-plugin would
require to add `spring-native-configuration` as a dependency to it.

This dependency contains both parts of the build-time infrastructure and
optional elements such as native hints annotations. Even if our goal is
to make this dependency completely optional, this commit includes it by
default with the plugin and makes it mandatory for native applications,
for now.

This will change in the future once we've refactored this dependency.

See gh-466