Move feature elements into buildtools
As one of the final pieces of the puzzle with 0.9.0 this
commit moves what the bti needs from the feature module
into the buildtools module. All that remains in the
feature module are the elements that depend upon GraalVM APIs.

The dependencies have been adjusted and flipped where relevant
(for example feature now depends on buildtools, not the other
way around).