Activate factory filtering and property checking in bti
With this commit the SpringFactoriesProcessor implementations that
the feature uses are now active at buildtime. This means the
ActuatorMetricsEliminator and BackgroundPreinitializerDeactivator
are turned on.

Also early condition processing at build time is turned on
so it is possible to check a few conditions that rely on
properties being set at buildtime, with full control over
which set of conditions are checked (based on specifying
prefixes for those included/excluded).