Rename the build-time transformation infrastructure to spring-aot
The Maven/Gradle plugins and the related shared module are not native
specific and some of the build-time transformations could be useful on
regular JVM as well. For example, the functional configuration
transformation could be useful for Spring Fu. Potentially, some of those
could be interesting for regular Spring Boot efficiency optimization
(to be verified during 0.9.x upcoming milestones).

As a consequence, we need a more general name than
spring-native-maven-plugin (same for Gradle) and the existing
spring-native-buildtools should be adapted as well. One candidate was
spring-build-transformation-maven-plugin but it was too long.
So I ended up with spring-aot-maven-plugin and spring-aot module names
that are short, meaningful (I think the key point of what we do here is
AOT transformations) and not unusual in Spring terminology (Spring AOP).

The signal that should enable native related transformations is for now
the presence of the spring-native dependency on the project classpath.

Closes gh-487