Performance improvements related to #380
This change includes enhancements to the configuration properties
analysis to try and get closer to how 0.8.2 behaved. 0.8.2 didn't
do the right thing for all types of configuration property usage
so now we are more correct it is not surprising we have to
pay a price for that. These changes try to minimize that impact.
There are also some hint adjustments that seem reasonable.

These changes also connect configuration property reflective
access to the build time property checking option. Using that
option can create a further saving in memory/image-size.