Be more consistent in access for types
Some code was deciding differently about access rather
than using the central decision maker algorithm. This
gave rise to a problem in the cloud sample where
an @Value field was not found in the application. I've
made this new spot go through the standard algorithm.
Might slightly inflate image sizes but consistency
and compatibility is improved.