Separate folders for generated sources and resources
Prior to this commit, generated content would be written to a single,
shared location; sources and resources would be organized later in
different locations.
This relies on a shared folder structure, which might not work for all
cases.

This commit splits the sources and resources locations in the AOT tools.