Use runtime classpath for compiling AOT sources
Prior to this commit, the AOT Maven plugin would use the default
classpath elements (the compile classpath) when compiling generated AOT
sources.
This could cause compilation errors as the generation phase is using the
runtime classpath and might generate code depending on runtime
dependnecies.

This commit ensures that the compiler for generated sources is using the
application runtime classpath.

Fixes gh-538