Fix parent ClassLoader in AOT generation
Prior to this commit, we would create a synthetic classloader using the
(test) runtime classpath of the application to perform AOT generation.
This synthetic ClassLoader would use the parent or the plugin
ClassLoader as its own parent.

On JDK9+, this would cause issues since classes that are not present in
the base module would not be present and class conditions would fail.

This commit ensures that we use the platform ClassLoader if we're on
JDK9+.

Fixes gh-566