Recreate generated AOT sources folder for each run
Prior to this commit, adding and removing dependencies during
developement could generate classes and leave them in place, but such
classes cannot be compiled anymore if a dependency has been removed.

This commit ensures that the generated sources folder by the AOT plugin
is re-created for each run.

Fixes gh-623