Use JDK base parent ClassLoader in AOT build context
Prior to this commit, we would use the parent of the current
`ClassLoader` as the parent for the AOT context `ClassLoader`. When
using Gradle, the parent `ClassLoader` contains the dependencies brought
by `buildSrc`.

This commit ensures that JDK8 environments are only based on the JDK
base `ClassLoader` as the parent.

Fixes gh-627