Add automated release process
This commit adds a new group to CI that allows to perform a release in
two steps. First, a staging build is pushed to `libs-staging-local`.

Then, a promotion job can be invoked once the bits have been verified.
For the moment, only a local promotion is performed that triggers the
automatic generation of the changelog.