Fix Spring Data custom repository fragment lookup.
Repository fragments following the pattern FragmentInterface | FragmentInterfaceImpl did not get discovered and registered correctly though annotated with @Component.
This commit introduces a Substitution for the CustomRepositoryImplementationDetector using the CandidateComponentsIndex directly. It also adds the missing type hint for the RepositoryFragment used to create the target repository and updates the data-elasticsearch example to cover the custom repository setup.
The alternative setup using the heritage configuration format (used in the data-mongodb) still works.