Build GraalVM images automatically
This commit watches the Graal github repository and trigger the build of
the springci/graalvm-ce images once a day if necessary.

Images are tagged with the milestone so that each branch can have
isolated configuration. Rather than putting the GraalVM version in the
image name, images are categorized as "stable" (the current maintenance
release) or "dev" (the current development branch).