Fix ConditionalOnBean processing in spring factories AOT
Prior to this commit, the AOT processing of spring factories would not
filter out auto-configurations if one of the types listed in
`@ConditionalOnBean` is present in the classpath.

In this case, we need to ensure that all types match the condition and
filter out the auto-configuration if it doesn't.

Fixes gh-718