Involve AOT sources only during image generation
Prior to this commit, the sources generated by Spring AOT (especially
the static variant of the Spring factories) would be involved at runtime
as soon as the spring-native dependency was on the classpath.

This would cause several limitations.

First, the development lifecycle could be limited by the build plugins
integration with the IDEs or the build tools themselves. In gh-758,
gh-675 and gh-593, we're seeing that using well-known build goals or
running the application from the IDE can cause problems.
Maybe the AOT sources were not generated yet. Or in general, this would
create confusion around generated sources and their use.

Also, the AOT generation could seriously limit existing use cases, such
as Spring Boot Devtools. Because devtools is not supported in native
images, a hard check was performed to prevent its usage. In gh-532, we
were trying to find a way to properly support that in all execution
modes, whereas this commit ensures that devtools is usable in most
cases.

This commit solves all those problems by introducing a runtime check in
our `SpringFactoriesLoader` override, only leveraging the generated
static spring factories if:

* the application is being compiled by GraalVM native
* or if the `springAot` System property has been set

While AOT sources are still generated and compiled at build time, they
are not being used by default at runtime on the JVM. The main use case
for using them at runtime is to diagnose and debug native image
generation problems - basically this gives developers a way to run the
application/its tests like in a mode optimized for the native image
generation.

Closes gh-532
Closes gh-675
Closes gh-758