Refine TypeProcessor - allow scanning TypeSystem without repeating filters.
Introduce TypeProcessor.filter and inspection depth to allow filtering candidate Types based on a single filter instead of repeated ones.
The inspection limit, defines the maximum number of hops allowed when inspecting types and their signatures.

Given a structure like below, and a level of 1 the callback would be invoked for Level0, SomeInterface
and Level1, but not for Level2.
When using level 0, the callback is only invoked for Level0.

class Level0 implements SomeInterface {
    Level1 level1;
}
class Level1 {
    Level2 level2;
}
class Level2 {
    // ...
}

Also introduce a DiscoveryContext that provides access to the Path that lead to a certain type. That way type registration callbacks can decide upon more information what to do.