Fix missing reflection config for mappings+startup endpoints
Prior to this commit, the "/actuator/mappings" and "/actuator/startup"
endpoints were not detected in the list of available actuator endpoints.

The endpoints are detected by looking at `@Endpoint` annotated beans and
their `@ReadOperation` `@WriteOperation` methods are registered as web
endpoints. In this case, the reflection information was missing for
those (the methods themselves, but also the return types to be
serialized).

This commit adds the missing reflection configuration to enable both
endpoints.

Fixes gh-851