Generalize code that notices if some types need reflective access
Before this change we would recognize non-components that had
@Autowired fields or methods but we wouldn't notice other
annotations like that. This change generalizes it to thinking
any annotation org.springframework.* probably needs reflective
access on these components.

At the same time, the previous code made all methods reflectable
if one needed this kind of access. With this more generalized
approach I've modified it to be more targeted and now
only the methods or fields in question will get reflective access.

The scheduled-tasks sample has been expanded with a test scenario
for this

Fixes #880