Handle generic parameter resolution
When a bean parameter contains a generic, we use the ObjectProvider
abstraction to define the resolvable type we're looking for but the
actual type signature is not retained. To help the compiler understand
the intent, a variable should be assigned.

This commit rework parameter resolution so that such cases lead to an
assignment within the instance supplier.