Add code snippet assertion utilities
Javapoet only fully replace imports when it writes a JavaFile. This
commit introduces a helper class that allows a test to generate an
arbitrary code snippet in the context of a method. The utility adds
special comment before and after the code snippet so that it can clean
the necessary noise around the code snippet under test.

This has the effect of providing a fully resolved code and expose the
imports in the future if we want to assert them. The test method can
also be customized so that we can easily copy/paste it from the error
message to debug it rather than looking at the code in text.