Use proper use case for build time parsing
This commit update the contributor to use the same context type as the
one that would be used at runtime. This is important as the
infrastructure that we don't handle yet one way or the other should
be registered so that we can generate the code for it.

Infrastructure beans that we handle already (typically configuration
class parsing at this time) are already excluded.