Document workaround for issue with nativeBuild plugin
This commit documents a workaround for a known compatibility issue
between the `org.springframework.experimental.aot` and the
`org.graalvm.buildtools.native` plugins.

The current design of the Spring AOT Gradle plugin contributes the
generated sources and configuration in a way that the GraalVM plugin
cannot be aware.

This workaround tells the native plugin about the additional locations
so that it can pick up the relevant sources.

Closes gh-873