Fork bootstrap generation process in build plugins
This commit updates the Maven and Gradle plugin to generate the
bootstrap code generation process in a forked process.
Prior to this commit, we would use the same JVM for code generation and
the actual build plugin - this could cause issues with class loading and
code generation since some Spring Framework infrastructure relies on the
thread context class loader for loading and checking classes.

Code generation is now handled by `BootstrapCodeGeneratorRunner`, with
all the required input given as command line arguments.

This commit also ensures that we don't require an `URLClassLoader` as
the default classloader - this is not the case as of JDK9.

Fixes gh-908