Add infrastructure to perform bean resolution
This commit adds a wrapper infrastructure used by the generated code to
resolve dependencies for methods, constructors, and fields.

BeanDefinitionRegistrar provides either a supplier or a function that
provides a dedicated context to resolve injection points. Such
callbacks transparently handle checked exception so that the generated
code does not have to handle it.

Using such an infrastructure also streamlines the registration of
bean definition with a `ResolvableType` as the new registrar provides
support for it.

See gh-909