Update code generation to honour injection points
This commit updates the code generation to account for injection
points. A BeanValueWriter now exposes the descriptor that it is
handling and, as a result, the default supplier creates such descriptor
rather than detecting the executable to use to instantiate the bean. As
a result, the custom bean value writer suppliers can now delegate to the
newly introduced SimpleBeanValueWriter.

BeanRegistrationGenerator uses the writer context to write the bean
registration code in a separate package if necessary. This makes the
registration code for a given bean self contained and easier to resonate
about. There is also a single default implementation now that the
BeanDefinitionRegistrar helper class accounts for both simple and
generic use cases.

ContextBootstrapGenerator registers an infrastructure bean that is
required to detect `@Qualifier` and `@Value` namely.

The generated code requires additional reflection for all injection
points. These are available in BeanDefinitionDescriptor and should be
added to the registry to make the generated code native-friendly.

See gh-909