Support runtime reflection configuration in Bootstrap Generator
This commit adds a wrapper to the existing `ReflectionDescriptor`
infrastructure; instead of being based on ASM, this wrapper uses
`java.lang.reflect`.

This allows the bootstrap code generator to add reflection configuration
when preparing the bean registry and generating the bootstrap source
code.

The generator now returns a `BootstrapGenerationResult` that contains
both the source files and the class descriptors. Both are set in the
context for further processing.

See gh-909

Co-authored-by: Stephane Nicoll <snicoll@vmware.com>