Add a type hint for SimpleClientHttpRequestFactory
The `connectTimeout` `readTimeout` and `bufferRequestBody` methods on
`RestTemplateBuilder` uses reflection to access underlying
`ClientHttpRequestFactory` implementation class.

Currently, type hints for `HttpComponentsClientHttpRequestFactory` and
`OkHttp3ClientHttpRequestFactory` exist, but the fallback request
factory, `SimpleClientHttpRequestFactory` hint doesn't exist.
Therefore, when `RestTemplate` doesn't detect either commons-client or
okhttp, it falls back to the `SimpleClientHttpRequestFactory` and the
reflective access fails to find the class.

This PR simply adds a hint to expose `SimpleClientHttpRequestFactory`,
so that above reflective methods on `RestTemplateBuilder` will have
access to it.

Fixes gh-954