Post-process bean definitions if necessary
This commit adds a strategy interface to post-process bean definitions
at build time. The main use case for now is Spring Data who doesn't
expose a resolved type for the factory bean used by repositories.

This commit makes sure that the target type refers to the repository
interface as well as the entity type so that injection happens
correctly at runtime.