Simplify "codeGenDebugPort" implementation in gradle plugin
Previously, the `codeGenDebugPort` implementation in gradle plugin was
not effectively using the `JavaExec` infrastructure.  This commit
updates to use the debug options from `JavaExec` to configure the
remote debugging.

For code simplification, remove the `springAot` DSL support and only
take `codeGenDebugPort` project property. This is because, in most
cases, remote debug is specified in the command line.

Also, adding documentation about `codeGenDebugPort` parameter.