Add support for bean value resolution in constructor injection
This commit reuses Framework's BeanDefinitionValueResolver to resolve
any constructor parameter prior to injection. This requires the
`BeanDefinition` to be available.

One of such possible parameter to resolve can be a BeanDefinition,
creating an "inner bean". This commit adds explicit support for those
using the `inner` factory type. Inner bean definitions are not available
from a `getBeanDefinition` call on a `ConfigurableBeanFactory`, so
the bean definition is stored once built and made available when the
element resolvers require it.

See gh-980