Write PropertyValues
This commit update the generator to write any PropertyValue that is
defined on a `BeanDefinition`. Such property could be a BeanDefinition
itself, which leads to a recursive call to the generator.

To handle this scenario, the generator now takes a function that allows
to initialize another instance for another bean definition. When doing
so, the `nesting` level is increased so that variable names don't clash
with the ones of the parent level.

See gh-980