Support AOT options in build plugins
Prior to this commit, the `BootstrapCodeGeneratorRunner` was implemented
as a simple class with a `main` method that is passed arguments as
positional parameters only. This would miss most AotOptions supported so
far and could be quite difficult to maintain.

This commit renames the runner to `GenerateBootstrap` and turns it into
a picocli CLI command with bound options and parameters. This now
supports all `remove*Support` options and some more.

The Maven and Gradle build plugins have been updated accordingly to call
the command with the expected arguments.

This commit also removes the need to push the classpath as a command
parameter and automatically extracts it using the `"java.class.path"`
system property.

Fixes gh-989