Configure Environment for AOT generation
Prior to this commit, the `ContextBootstrapContributor` would prepare an
empty `Environment` and only apply some defaults to it. This lead to
several sample applications failing as properties provided by the
application (in `application.properties`) would not be considered.

As a consequence, all `@ConditionalOnProperty` on `*.enabled`
configuration keys would be discarded at build time if not enabled by
default in Spring Boot.

This commit loads a minimal infrastructure and prepares an `Environment`
that partially loads the application configuration at build time.

There are many limitations with this current approach:

* this doesn't support many customizations on the `SpringApplication`
* this doesn't load `Bootstrapper` so will fail with configuration
  trying to resolve properties using `Bootstrappers`, like Eureka.

Fixes gh-992