Update Gradle options for AOT debugging
Prior to this commit, the Gradle AOT task would use a project property
to trigger remote debugging of the forked process for the AOT
generation. Project properties are unusual for this use case and this
can be challenging in case of a multi-module application.

This commit switches those to System properties: one to enable debugging
on the default 5005 port, and another one for customizing the port.