Allow to provide the BeanRegistrationWriter to use
This commit removes BeanValueWriterSupplier and replaces it with a
strategy interface to provide the BeanRegistrationWriter instead. This
way, specific cases can not only replace how the instance supplier is
generated, but also how the bean registration itself is described.

A new BeanRegistrationWriterOptions is introduced to abstract some of
the options that we want to expose.

See gh-1013