Add support for InstanceAware callback
This commit adds support for a specific configuration class callback
that is required at runtime for the actual bean instances to be
created correctly.

The InstanceAware callback requires the annotation metadata of the
configuration class that imported it, so that the original annotations
can be read, and custom annotation attributes can be tuned to configure
the infrastructure.

This requires to make a link between the current configuration class
and the configuration class that imported it. Such a link is written
in a method that is used to create a specific bean responsible to
honor the callback at runtime.

Closes gh-996