Select appropriate constructor for bean instantiation
Prior to this commit, the `BeanInstanceExecutableSupplier` used the
resolved factory method or constructor method information to find the
Executable to be called for instantiation.

If no info is available at the bean definition level, this was trying to
detect a suitable constructor - by applying a simple resolution
algorithm that detects a single constructor or a constructor annotated
with @Autowired.

This commit adds a simple algorithm for finding a suitable constructor
that fits the given constructor arguments.

Fixes gh-1015