Use plain constructor calls in BeanValueWriter
Prior to this commit, there were two variants of
`BeanDefinitionRegistrar#instanceSupplier` - one that takes a
`SmartFunction` and another that takes a `SmartSupplier`.

In some cases, the `DefaultBeanValueWriter` used a method handle like
`SampleBean::new` as an argument there. If the `SampleBean` has both a
no-arg constructor and a single-arg constructor, the compiler cannot
infer which variant needs to be called and fails.

This commit transforms the source code generated from `SampleBean::new`
to `() -> new SampleBean()` to avoid confusions.

Fixes gh-1018