Simplify bean registration structure
This commit extracts the responsibility of registering reflection and
resources in a separate class, based on the BeanInstanceDescriptor.

This made apparent that the current arrangement with an BeanValueWriter
and a BeanRegistrationWriter could be simplified with a single
interface that is responsible to write the registration and provide
the descriptor that it handles.

See gh-1023