Migrate spring.factories key related inferencing into Aot
With this change the 'old inferencer' is no longer chasing down
configuration from spring.factories entries (like the
EnableAutoConfiguration values). Instead the 'lookup of hints
related to configuration' is pushed to being done in Aot processing.

Because the current Aot processing doesn't have a way to pass
some of the stuff coming from hints through its registry construct
some samples that rely on hints adding resources/proxies etc may fail
but I wanted to put this in as a first step for review of where this
code should live. If it is vaguely in the right place, augmenting
the registry is straightforward. If it is in the wrong place, maybe
it can move to a place not reliant on the registry as a passthrough.