Configure multiple sourceSet outputs for AOT generation
Prior to this commit, the AOT plugins would only consider a single path
as a location for the application compiled classes. While this is true
for Maven builds, Gradle builds separate outputs for each sourceSet (in
our case, one for Java and one for Kotlin).

This commit ensures that all sourceSets are considered for classes
outputs. This also applies to scanning for the main application class.

Fixes gh-1019