Adapt ApplicationContext and Environment if necessary
Previously, if the configuration of the application contains a custom
web application type, it was ignored and both the context and the
environment were as detected by the classpath. This led to problems
as the generated code is for an environment, while the runtime one
was using another.

This commit makes sure that the value is taken into account and the
environment is adapted. It also harmonizes the behavior with the latest
improvements in Spring Boot 2.6

Closes gh-1053