Invoke init and destroy callbacks
This commit rework the original proposal by writing the init and
destroy methods and using them to instantiate a dedicated bean post
processor that is similar to the original one used at runtime.

Both method names defined on the bean definition itself as well as
externally managed ones are supported. If the bean name is '(inferred)'
similar logic is applied at build time.

Closes gh-1048
See gh-1059