Find lifecycle method on declared bean type if necessary
This commit improves `InitDestroyBeanPostProcessor` to fallback on the
declared type of the bean if it's actual implementation class does not
provide access to the lifecycle method.

This happens on GraalVM when the hint is declared on an interface as
framework's algorithm only consider superclasses and GraalVM won't
expose the method if it has not been configured on the type explicitly.

Closes gh-1121