Add infrastructure to generate test contexts
This commit adds the necessary infrastructure to generate the context
from a test class. This essentially reuses piece of the Test Context
Framework, retrieving the boostrapper, creating the
MergedContextConfiguration and listing all the contexts that are
required for a module.

Each context is generated in its own class with a main class that
defines the mapping between each test and its context. An AOT specific
context loader delegate is also available that looks to the mapping and
use the generated context rather than parsing it at runtime.

Closes gh-1162