Refactor AOT Gradle plugin
This commit refactors the AOT Gradle build plugin and changes important
behavior.

First, this commit upgrades to the GraalVM Native Build Tools 0.9.5
version. This version provides a new Gradle extension and new
configuration options.
With that, we're also now depending on this plugin and applying
directly. The AOT build plugin also configures automatically the
`nativeCompile` task to add the generated sources to the build
classpath.

Also, this plugin changes the general structure of the build plugin to
offer an `aotMain` project configuration that holds the generated
sources and exports a compiled jar of generated sources as an output.
This means that tasks depending on the `aotMain` configuration (and thus
on the `generateAot` task) will now get the JAR of compiled generated
sources and add it to their classpath. This JAR is produced with an
`aot` classifier and has the same name as the main application JAR.

Closes gh-1031
Closes gh-922