Fix test script for Gradle samples
This commit fixes the new location of the GraalVM configuration for
Gradle projects, `aotMain` instead of `aot`.
This also ensures that the proper jar is used for testing, since we're
generating several variants for gradle projects (the "plain" jar which
is not repackaged, the "aot" jar with generated sources and the main
repackaged jar).