Add support for Actuator management context
This commit adds support for the separate application context that
Spring Boot creates when a management port is set. In AOT mode, the
context is processed at build time, exactly as the maine one is, and
the ManagementContextFactory implementation is replaced by an
AOT-specific one that loads the context from the generated code.

Closes gh-1146