Upgrade to Native Build Tools 0.9.7.1
This commit upgrades the spring-aot-gradle-plugin and all Maven based
sample apps to Native Build Tools (NBT) 0.9.7.1

NBT 0.9.6 requires JUnit Platform 1.8 (JUnit 5.8) due to its dependency
on the UniqueIdTrackingListener in the junit-platform-launcher
artifact. The Maven sample apps are currently based on Spring Boot 2.6
RC1 which uses JUnit 5.8.1 (JUnit Platform 1.8.1). Thus, there is no
conflict between the JUnit version used in the sample apps and the
version used by NBT.

However, there is an issue regarding Maven Surefire.

Maven Surefire 3.0.0-M4 introduced support for automatically aligning
the JUnit Platform version used by Surefire with the version needed by
the user's configured version of JUnit Jupiter. Since we are not using
Surefire 3.0 milestones, the parent POM for the Maven sample apps has
to explicitly declare a dependency on the correct version of the
junit-platform-launcher in order to make the UniqueIdTrackingListener
available for Native Build Tools 0.9.6 and higher. The actual version
of the junit-platform-launcher is managed by the junit-bom that is
imported by the spring-boot-starter-parent.

Closes gh-1202