Improve detection of health indicator types
This commit handles the composite pattern used in Spring Boot that
exposes either the type directly or the composite if more than one
instance of the indicator is required. Because a single method is used
for both, the return type does not provide the actual indicator type
implementation.

By checking the configurations that are used to expose such a bean, we
can resolve the generic type that provides the indicator type and
register it that way instead.

Closes gh-1183