Invoke BeanDefinitionRegistryPostProcessor at build time
This commit improves the parsing of the bean factory at build time to
invoke BeanDefinitionRegistryPostProcessor defined as beans. When such
a bean has been processed at build time, it is removed from the list of
bean definitions to prevent it from contributing to the registry again
at runtime. To limit the side effect, only beans registered with role
infrastructure are removed this way.

Closes gh-1213