Prevent invalid root directory errors
Before this commit, MainClassFinder and variants were
used for all classpath entries without differentiating
between directories or JAR, while MainClassFinder
provides distinct methods for "File rootDirectory"
versus "JarFile jarFile".

This commit tests if the related file is a directory
before invoking methods like findSingleMainClass.

If needed, it could be refined to scan JAR files as well,
but it could potentially have side effects, so let's be
defensive and not change that behavior for now.

Closes gh-1209