Ensure CLR tests fail in native image if ApplicationContext doesn't load
Prior to this commit, the CommandLineRunner (CLR) sample applications
each had a CommandlinerunnerApplicationTests test class that did not
assert anything.

Consequently, since the switch to AOT testing support, the tests have
been passing even though the ApplicationContext was not loaded in the
native image, which results in false positives in the build.

This commit addresses this issue by modifying each of those test classes
to implement ApplicationContextAware and then check that the supplied
context is not null.

This is intentionally done via the ApplicationContextAware callback
instead of autowiring, since we will address proper auto-wiring support
in the near future.

Closes gh-1216